Fixed empty chunk issue, still unexpected errors

+++ src/common.hh
@@ -10,8 +10,8 @@
 #include <map>
 
 /// \brief Recherche ou ajout de chaine dans le dictionnaire
-const std::pair<bool, std::uint16_t>
-dico(std::map<std::pair<std::uint16_t, std::uint8_t>, std::uint16_t> &,
-     uint16_t, uint8_t);
+const std::pair<bool, std::uint32_t>
+dico(std::map<std::pair<std::uint32_t, std::uint8_t>, std::uint32_t> &,
+     uint32_t, uint8_t);
 
 #endif /* LZW_SRC_COMMON_H_ */
